I updated the benchmark data because it did not capture the creation of the implemenation type MoveInstruction, StateWaypoint, JointWaypoint and CartesianWaypoint. It does show that adding the boost::uuids::uuid to the MoveInstruction increase the creation time from 155ns to 900ns. This still a very small time where if you were to create 1000 instruction the time would to create all move instructions would be 0.0009s vs 0.000155s and since this does not impact copying which is what occurs during the planning pipeline I feel this is acceptable.

Before:

``` bash
-----------------------------------------------------------------------------
Benchmark                                   Time             CPU   Iterations
-----------------------------------------------------------------------------
BM_InstructionPolyCreation              0.000 ns        0.000 ns   1000000000
BM_WaypointPolyCreation                 0.000 ns        0.000 ns   1000000000
BM_MoveInstructionCreation                155 ns          155 ns      4427227
BM_StateWaypointCreation                 52.3 ns         52.3 ns     13160799
BM_JointWaypointCreation                 46.6 ns         46.6 ns     14589245
BM_CartesianWaypointCreation             10.7 ns         10.7 ns     65025243
BM_CompositeInstructionCreation          49.8 ns         49.8 ns     13885873
BM_ProgramCreation                      26792 ns        26790 ns        25928
BM_InstructionPolyCopy                   62.3 ns         62.3 ns     10998026
BM_WaypointPolyCopy                      19.6 ns         19.6 ns     35652538
BM_CompositeInstructionCopy              6128 ns         6128 ns       118155
BM_InstructionPolyAssign                 62.6 ns         62.6 ns     11250108
BM_WaypointPolyAssign                    19.7 ns         19.7 ns     35535057
BM_CompositeInstructionAssign            5806 ns         5806 ns       119602
BM_InstructionPolyCast                   1.96 ns         1.96 ns    355923076
BM_WaypointPolyCast                      1.75 ns         1.75 ns    400402308
BM_InstructionPolyAccess                 2.18 ns         2.18 ns    320238369
BM_WaypointPolyAccess                   0.218 ns        0.218 ns   1000000000
BM_VectorStateWaypointPolyCreation     153191 ns       153190 ns         4612
BM_VectorStateWaypointUPtrCreation     144173 ns       144171 ns         4875
BM_VectorStateWaypointPolyCopy          86024 ns        86023 ns         8134
BM_VectorStateWaypointUPtrCopy          86456 ns        86455 ns         8058
```

After:

``` bash
-----------------------------------------------------------------------------
Benchmark                                   Time             CPU   Iterations
-----------------------------------------------------------------------------
BM_InstructionPolyCreation              0.000 ns        0.000 ns   1000000000
BM_WaypointPolyCreation                 0.000 ns        0.000 ns   1000000000
BM_MoveInstructionCreation                900 ns          900 ns       771254
BM_StateWaypointCreation                 54.6 ns         54.6 ns     13039459
BM_JointWaypointCreation                 47.6 ns         47.6 ns     14743474
BM_CartesianWaypointCreation             10.6 ns         10.6 ns     66011077
BM_CompositeInstructionCreation          51.1 ns         51.1 ns     13678986
BM_ProgramCreation                      34201 ns        34200 ns        20446
BM_InstructionPolyCopy                   66.5 ns         66.5 ns     10512135
BM_WaypointPolyCopy                      20.4 ns         20.4 ns     34139015
BM_CompositeInstructionCopy              5820 ns         5820 ns       120516
BM_InstructionPolyAssign                 65.3 ns         65.3 ns     10846574
BM_WaypointPolyAssign                    21.3 ns         21.3 ns     33852431
BM_CompositeInstructionAssign            5905 ns         5905 ns       118460
BM_InstructionPolyCast                   1.77 ns         1.77 ns    397620001
BM_WaypointPolyCast                      1.79 ns         1.79 ns    394245538
BM_InstructionPolyAccess                 2.21 ns         2.21 ns    315709956
BM_WaypointPolyAccess                   0.219 ns        0.219 ns   1000000000
BM_VectorStateWaypointPolyCreation     151456 ns       151453 ns         4644
BM_VectorStateWaypointUPtrCreation     145443 ns       145440 ns         4800
BM_VectorStateWaypointPolyCopy          86364 ns        86364 ns         8097
BM_VectorStateWaypointUPtrCopy          86519 ns        86518 ns         8082
```